requirements.txt
amqp
asgiref
astroid
billiard
celery
certifi
chardet
click
click-didyoumean
click-repl
coverage
dj-database-url
Django
django-heroku
djangorestframework
gunicorn
idna
isort
kombu
lazy-object-proxy
mccabe
Pillow
prompt-toolkit
protobuf
psycopg2
pylint
python-decouple
pytz
requests
six
sqlparse
toml
urllib3
vine
wcwidth
whitenoise
wrapt
